requirements.txt
PySimpleSOAP==1.10
suds-jurko==0.6
